% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{atomic_density}
\alias{atomic_density}
\title{Atomic Density}
\usage{
atomic_density(rho, A)
}
\arguments{
\item{rho}{Material density (g/cm³)}

\item{A}{Atomic mass (g/mol)}
}
\value{
Atomic density (atoms/cm³)
}
\description{
Calculates number density n in σ = μ / n
}
