% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{attenuation_law}
\alias{attenuation_law}
\title{Exponential Attenuation Law}
\usage{
attenuation_law(I0, mu, x)
}
\arguments{
\item{I0}{Initial intensity}

\item{mu}{Attenuation coefficient (cm⁻¹)}

\item{x}{Distance traveled (cm)}
}
\value{
Attenuated intensity
}
\description{
Relates attenuation coefficient μ to microscopic cross section
}
