% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing_tools.R
\name{attribution}
\alias{attribution}
\title{Multi-Touch Attribution Model}
\usage{
attribution(touchpoints, model = "linear", position_weights = c(0.4, 0.2, 0.4))
}
\arguments{
\item{touchpoints}{Vector of touchpoint numbers (per channel)}

\item{model}{Attribution type: "linear", "time_decay", "position"}

\item{position_weights}{For "position": c(first, middle, last)}
}
\value{
Vector of credits attributed to each touchpoint
}
\description{
Attributes conversion credit to different touchpoints
}
