% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.R
\name{barn_to_cm2}
\alias{barn_to_cm2}
\title{Conversion barn → cm2}
\usage{
barn_to_cm2(b)
}
\arguments{
\item{b}{Cross section in barn}
}
\value{
Cross section in cm2
}
\description{
Converts cross section from barn to cm²
}
\details{
1 barn = 10e-24 cm2
}
