% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.R
\name{barn_to_m2}
\alias{barn_to_m2}
\title{Conversion barn → m2}
\usage{
barn_to_m2(b)
}
\arguments{
\item{b}{Cross section in barn}
}
\value{
Cross section in m2
}
\description{
Converts cross section from barn to square meters
}
\details{
1 barn = 10exp(-28) m2 = 10exp(-24) cm2
}
\examples{
barn_to_m2(1)      # 1e-28 m2
barn_to_m2(100)    # 1e-26 m2
}
