% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{breit_wigner}
\alias{breit_wigner}
\title{Reaction Cross Section (Breit-Wigner model)}
\usage{
breit_wigner(E, E0, Gamma, sigma0)
}
\arguments{
\item{E}{Energy(ies) in MeV}

\item{E0}{Resonance energy in MeV}

\item{Gamma}{Resonance width in MeV}

\item{sigma0}{Peak amplitude in barn}
}
\value{
Cross section σ(E) in barn
}
\description{
Models nuclear resonance peaks
}
\details{
Breit-Wigner formula:
  σ(E) = σ_0 · (Γ²/4) / [(E - E_0)² + (Γ²/4)]
}
