% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{capture_cross_section}
\alias{capture_cross_section}
\title{Capture Cross Section}
\usage{
capture_cross_section(E, a, b)
}
\arguments{
\item{E}{Energy(ies) in eV}

\item{a}{Coefficient in barn·√eV}

\item{b}{Coefficient in barn/eV}
}
\value{
Capture cross section σ_capture(E) in barn
}
\description{
Simple model for neutron capture
}
\details{
σ(E) = a/√E + b·E
  First term: thermal capture (1/v law)
  Second term: resonance capture
}
