% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.R
\name{cm2_to_barn}
\alias{cm2_to_barn}
\title{Conversion cm2 → barn}
\usage{
cm2_to_barn(cm)
}
\arguments{
\item{cm}{Cross section in cm2}
}
\value{
Cross section in barn
}
\description{
Converts cross section from cm2 to barn
}
\details{
1 cm2 = 10e24 barn
}
