% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing_tools.R
\name{conversion_rate}
\alias{conversion_rate}
\title{Conversion Rate}
\usage{
conversion_rate(conversions, impressions, pct = FALSE)
}
\arguments{
\item{conversions}{Number of conversions}

\item{impressions}{Number of impressions}

\item{pct}{If TRUE, returns as percentage}
}
\value{
Conversion rate (0-1 or percentage)
}
\description{
Calculates the simple conversion rate
}
