% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{cross_section_error}
\alias{cross_section_error}
\title{Error Propagation on Cross Section}
\usage{
cross_section_error(dN, N0, n, dx)
}
\arguments{
\item{dN}{Number of detected reactions}

\item{N0}{Number of incident particles}

\item{n}{Target density (atoms/cm³)}

\item{dx}{Interval width}
}
\value{
List with: sigma (value), error (uncertainty)
}
\description{
Calculates cross section with statistical uncertainty
}
