% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.R
\name{derivative}
\alias{derivative}
\title{Numerical Derivative}
\usage{
derivative(x, y, method = "central")
}
\arguments{
\item{x}{Vector of x-values}

\item{y}{Vector of y-values}

\item{method}{Method: "forward", "backward", "central"}
}
\value{
Vector of derivatives dy/dx
}
\description{
Calculates derivative of sampled function
}
\details{
- "forward": (y[i+1] - y[i]) / (x[i+1] - x[i])
  - "backward": (y[i] - y[i-1]) / (x[i] - x[i-1])
  - "central": (y[i+1] - y[i-1]) / (x[i+1] - x[i-1]) (more accurate)
}
