% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{differential_cross_section}
\alias{differential_cross_section}
\title{Differential Cross Section Calculation}
\usage{
differential_cross_section(dN, N0, n, dx)
}
\arguments{
\item{dN}{Number of detected reactions}

\item{N0}{Number of incident particles}

\item{n}{Target density (atoms/cm³)}

\item{dx}{Interval width (angle, energy, or distance)}
}
\value{
Differential cross section
}
\description{
Analyzes a small angular, energetic, or spatial interval
}
