% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.R
\name{find_peaks}
\alias{find_peaks}
\title{Peak Finding}
\usage{
find_peaks(x, threshold = 0.1, min_distance = 1)
}
\arguments{
\item{x}{Vector of values}

\item{threshold}{Relative threshold (0-1) for peak detection}

\item{min_distance}{Minimum distance between two peaks}
}
\value{
Indices of detected peaks
}
\description{
Detects local maxima in a signal
}
\examples{
E <- seq(0, 10, length.out = 500)
sigma <- breit_wigner(E, 3, 0.2, 10) +
         breit_wigner(E, 6, 0.3, 8)
peaks <- find_peaks(sigma, threshold = 0.3)
}
