% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{isotropic_scattering}
\alias{isotropic_scattering}
\title{Isotropic Scattering}
\usage{
isotropic_scattering(sigma_total)
}
\arguments{
\item{sigma_total}{Total cross section (barn)}
}
\value{
Differential cross section (barn/sr)
}
\description{
Converts total cross section to isotropic differential
}
