% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{legendre_scattering_fast}
\alias{legendre_scattering_fast}
\title{Anisotropic Scattering (optimized vectorized version)}
\usage{
legendre_scattering_fast(theta, coeff)
}
\arguments{
\item{theta}{Angle(s) in radians}

\item{coeff}{Coefficient vector [a_0, a_1, a_2, ..., a_L]}
}
\value{
Angular cross section σ(θ) in barn/sr
}
\description{
Fast version using matrix pre-computation
}
\details{
This version is 10-20x faster for large datasets
}
