% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing_tools.R
\name{ltv}
\alias{ltv}
\title{Customer Lifetime Value (LTV)}
\usage{
ltv(
  avg_purchase_revenue,
  annual_purchase_freq,
  customer_lifespan,
  discount_rate = 0.1
)
}
\arguments{
\item{avg_purchase_revenue}{Average revenue per purchase}

\item{annual_purchase_freq}{Purchase frequency per year}

\item{customer_lifespan}{Customer lifespan (years)}

\item{discount_rate}{Annual discount rate (0-1)}
}
\value{
Discounted LTV
}
\description{
Estimates customer lifetime value
}
