% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.R
\name{m2_to_barn}
\alias{m2_to_barn}
\title{Conversion m2 → barn}
\usage{
m2_to_barn(m)
}
\arguments{
\item{m}{Cross section in m2}
}
\value{
Cross section in barn
}
\description{
Converts cross section from square meters to barn
}
\details{
1 m2 = 10e28 barn
}
\examples{
m2_to_barn(1e-28)  # 1 barn
m2_to_barn(1e-24)  # 10000 barn
}
