% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{micro_cross_section}
\alias{micro_cross_section}
\title{Microscopic from Macroscopic}
\usage{
micro_cross_section(mu, n)
}
\arguments{
\item{mu}{Macroscopic cross section (cm⁻¹)}

\item{n}{Atomic density (atoms/cm³)}
}
\value{
Microscopic cross section
}
\description{
Microscopic from Macroscopic
}
