% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing_tools.R
\name{optimize_budget}
\alias{optimize_budget}
\title{Optimal Budget Allocation (2 channels)}
\usage{
optimize_budget(total_budget, params_A, params_B, resolution = 100)
}
\arguments{
\item{total_budget}{Total available budget}

\item{params_A}{Vector c(Emax_A, EC50_A, n_A) for channel A}

\item{params_B}{Vector c(Emax_B, EC50_B, n_B) for channel B}

\item{resolution}{Number of test points (default: 100)}
}
\value{
List with: budget_A, budget_B, total_response
}
\description{
Finds optimal budget split between two channels
}
