% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_tools.R
\name{plot_angular_distribution}
\alias{plot_angular_distribution}
\title{Angular Differential Cross-Section Plot}
\usage{
plot_angular_distribution(
  theta,
  sigma_theta,
  type = "polar",
  main = "Angular distribution"
)
}
\arguments{
\item{theta}{Numeric vector of angles (degrees).}

\item{sigma_theta}{Numeric vector of differential cross-section values (barn/sr).}

\item{type}{Either `"polar"` or `"cartesian"`.}

\item{main}{Title of the plot.}
}
\value{
A polar or Cartesian angular distribution plot.
}
\description{
Visualizes a differential cross section \eqn{\frac{d\sigma}{d\Omega}(\theta)} either
in polar coordinates or in standard Cartesian coordinates.
}
\examples{
theta <- seq(0, 360, by = 1)
sigma <- 3 + sin(theta * pi/180)
plot_angular_distribution(theta, sigma, type = "polar")

}
