% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_tools.R
\name{plot_heatmap_sigma}
\alias{plot_heatmap_sigma}
\title{2D Cross-Section Heatmap}
\usage{
plot_heatmap_sigma(E, theta, sigma_matrix, main = "2D cross section")
}
\arguments{
\item{E}{Vector of energies (x-axis).}

\item{theta}{Vector of angles (y-axis).}

\item{sigma_matrix}{Matrix of cross-section values.}

\item{main}{Title of the plot.}
}
\value{
A 2D heatmap with contour lines.
}
\description{
Generates a color heatmap of a cross section depending simultaneously on
energy and angle: \eqn{\sigma(E, \theta)}.
}
