% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketing_tools.R
\name{romi}
\alias{romi}
\title{Return On Marketing Investment (ROMI)}
\usage{
romi(revenue, campaign_cost, pct = FALSE)
}
\arguments{
\item{revenue}{Revenue generated by the campaign}

\item{campaign_cost}{Campaign cost}

\item{pct}{If TRUE, returns as percentage}
}
\value{
ROMI (ratio or %)
}
\description{
Calculates marketing ROI
}
