% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{rutherford_scattering}
\alias{rutherford_scattering}
\title{Rutherford Scattering}
\usage{
rutherford_scattering(theta, Z1, Z2, E)
}
\arguments{
\item{theta}{Scattering angle(s) in degrees}

\item{Z1}{Atomic number of projectile}

\item{Z2}{Atomic number of target}

\item{E}{Projectile energy in MeV}
}
\value{
Differential cross section in barn/sr
}
\description{
Flagship function for differential cross section
}
\details{
Formula: dσ/dΩ = (k²) / sin⁴(θ/2), where k = Z₁Z₂e²/(4E)
}
