% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.R
\name{smooth_data}
\alias{smooth_data}
\title{Smooth Noisy Data}
\usage{
smooth_data(x, window_size = 5, method = "moving_avg")
}
\arguments{
\item{x}{Vector to smooth}

\item{window_size}{Smoothing window size (odd number)}

\item{method}{Method: "moving_avg" or "savitzky_golay"}
}
\value{
Smoothed vector
}
\description{
Applies smoothing filter to data
}
\details{
- "moving_avg": simple moving average
  - "savitzky_golay": polynomial filter (preserves peaks better)
}
