% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitaires.R
\name{spectrum_stats}
\alias{spectrum_stats}
\title{Spectrum Descriptive Statistics}
\usage{
spectrum_stats(E, sigma)
}
\arguments{
\item{E}{Energy vector}

\item{sigma}{Cross section vector}
}
\value{
List with: mean, median, max, max_E, integral, FWHM
}
\description{
Calculates statistics of a cross section
}
