% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{total_cross_section}
\alias{total_cross_section}
\title{Total Cross Section (numerical integration)}
\usage{
total_cross_section(angle, sigma_angle)
}
\arguments{
\item{angle}{Vector of angles or energies}

\item{sigma_angle}{Vector of corresponding cross sections}
}
\value{
Integrated total cross section
}
\description{
Integrates over an angular or energy domain
}
