% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_tools.R
\name{trapz}
\alias{trapz}
\title{Trapezoidal Integration (internal implementation)}
\usage{
trapz(x, y)
}
\arguments{
\item{x}{Vector of x-values (abscissa)}

\item{y}{Vector of y-values (ordinate)}
}
\value{
Approximate integral
}
\description{
Trapezoidal rule for numerical integration
}
